Updated OddOccurences and PermMissingElem code for better coding practices.

+++ TimeComplexity/PermMissingElem.cpp
@@ -12,16 +12,16 @@ namespace CodilityTasks
 
 int CodilityTasks::PermMissingElem::solution(std::vector<int> &A)
 {
-	long long n = A.size() + 1;
-	long long expectedSum = (n * (n + 1)) / 2;
+	int n = A.size() + 1;
+	long expectedSum = (n * (n + 1)) / 2;
 
-	long long arraySum = 0;
+	long arraySum = 0;
 	for (int num : A)
 	{
 		arraySum += num;
 	}
 
-	long long missingElement = expectedSum - arraySum;
+	long missingElement = expectedSum - arraySum;
 
 	return static_cast<int>(missingElement);
 }
